requirements.txt
requests
python-dotenv
pandas
pyarrow
numpy
matplotlib
seaborn
statsmodels
pmdarima
scikit-learn
torch
joblib
